Remove matrix inverse code - not needed anymore
