+++ lib/matrix.h
@@ -103,7 +103,6 @@ class Matrix
 };
 
 #include "matrix_free_function.hpp"
-#include "matrix_inverse.hpp"
 #include "matrix_operator.hpp"
 
 // Matrix:: member functions
